move people list cubit state to separate file
