+++ lib/ui/widgets/people_widget/popular_people_list_cubit.dart
@@ -2,6 +2,7 @@
 import 'dart:async';
 
 // Flutter imports:
+import 'package:comics_db_app/ui/widgets/people_widget/components/people_list_cubit_state.dart';
 import 'package:flutter/material.dart';
 
 // Package imports:
@@ -14,33 +15,6 @@ import 'package:comics_db_app/ui/navigation/main_navigation.dart';
 import 'package:comics_db_app/ui/widgets/people_widget/components/people_list_data.dart';
 
 // TODO replace cubit state
-class PeopleListCubitState {
-  final List<PeopleListData> people;
-  final String localeTag;
-
-  PeopleListCubitState({required this.people, required this.localeTag});
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is PeopleListCubitState &&
-          runtimeType == other.runtimeType &&
-          people == other.people &&
-          localeTag == other.localeTag;
-
-  @override
-  int get hashCode => people.hashCode ^ localeTag.hashCode;
-
-  PeopleListCubitState copyWith({
-    List<PeopleListData>? people,
-    String? localeTag,
-  }) {
-    return PeopleListCubitState(
-      people: people ?? this.people,
-      localeTag: localeTag ?? this.localeTag,
-    );
-  }
-}
 
 class PeopleListCubit extends Cubit<PeopleListCubitState> {
   final PeopleListBloc peopleListBloc;
